Organize files to test in a folder
