+++ app_code/book.rb
