+++ app_code/capitalize_decorator.rb
