+++ app_code/nameable.rb
