+++ app_code/rental.rb
