+++ app_code/trimmer_decorator.rb
